Wrap ShoppingCartItem in a Status
Now we just need to check this one Status to know if all went well.
There is a problem though: This doesn't scale. At all. Just imagine what
this would look like if we add a parameter to the constructor of
ShoppingCartItem. We would need 3 levels deep worth of whens.

If only we could make this whole process a bit more sequential...